+++ upsetplot/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.6.dev1'
+__version__ = '0.7.dev1'
 
 import os